+++ lib/aspera/cli/formatter.rb
@@ -98,11 +98,12 @@ module Aspera
       CSV_RECORD_SEPARATOR = "\n"
       CSV_FIELD_SEPARATOR = ','
       # supported output formats
-      DISPLAY_FORMATS = %i[text nagios ruby json jsonpp yaml table csv image].freeze
+      DISPLAY_FORMATS = %i[text nagios ruby json jsonpp yaml table multi csv image].freeze
       # user output levels
       DISPLAY_LEVELS = %i[info data error].freeze
+      FIELD_VALUE_HEADINGS = %i[key value].freeze
 
-      private_constant :DISPLAY_FORMATS, :DISPLAY_LEVELS, :CSV_RECORD_SEPARATOR, :CSV_FIELD_SEPARATOR
+      private_constant :DISPLAY_FORMATS, :DISPLAY_LEVELS, :CSV_RECORD_SEPARATOR, :CSV_FIELD_SEPARATOR, :FIELD_VALUE_HEADINGS
       # prefix to display error messages in user messages (terminal)
       ERROR_FLASH = 'ERROR:'.bg_red.gray.blink.freeze
       WARNING_FLASH = 'WARNING:'.bg_brown.black.blink.freeze
@@ -326,16 +327,16 @@ module Aspera
           display_message(:info, special_format('empty')) if @options[:format].eql?(:table)
           return
         end
+        # if table has only one element, and only one field, display the value
         if object_array.length == 1 && fields.length == 1
           display_message(:data, object_array.first[fields.first])
           return
         end
         # Special case if only one row (it could be object_list or single_object)
         if @options[:transpose_single] && object_array.length == 1
-          new_columns = %i[key value]
           single = object_array.first
-          object_array = fields.map { |i| new_columns.zip([i, single[i]]).to_h }
-          fields = new_columns
+          object_array = fields.map { |i| FIELD_VALUE_HEADINGS.zip([i, single[i]]).to_h }
+          fields = FIELD_VALUE_HEADINGS
         end
         Log.log.debug{Log.dump(:object_array, object_array)}
         # convert data to string, and keep only display fields
@@ -348,8 +349,18 @@ module Aspera
             headings:  fields,
             rows:      final_table_rows,
             style:     @options[:table_style]&.symbolize_keys))
+        when :multi
+          final_table_rows.each do |row|
+            Log.log.debug{Log.dump(:row, row)}
+            display_message(:data, Terminal::Table.new(
+              headings:  FIELD_VALUE_HEADINGS,
+              rows:      fields.zip(row),
+              style:     @options[:table_style]&.symbolize_keys))
+          end
         when :csv
           display_message(:data, final_table_rows.map{|t| t.join(CSV_FIELD_SEPARATOR)}.join(CSV_RECORD_SEPARATOR))
+        else
+          raise "not expected: #{@options[:format]}"
         end
       end
 
@@ -420,7 +431,7 @@ module Aspera
           end
           raise "not url: #{url.class} #{url}" unless url.is_a?(String)
           display_message(:data, status_image(url))
-        when :table, :csv
+        when :table, :csv, :multi
           case type
           when :config_over
             display_table(Flattener.new(self).config_over(data), CONF_OVERVIEW_KEYS)
@@ -452,6 +463,8 @@ module Aspera
           else
             raise "unknown data type: #{type}"
           end
+        else
+          raise "not expected: #{@options[:format]}"
         end
       end
     end